+++ tests/models_o2o_2.py
@@ -1,0 +1,9 @@
+"""
+This is the testing Models — Bad on_delete parameter
+"""
+from tortoise import fields
+from tortoise.models import Model
+
+
+class One(Model):
+    tournament = fields.OneToOneField("models.Two", on_delete="WABOOM")