+++ tortoise/backends/sqlite/client.py
@@ -44,6 +44,7 @@ class SqliteClient(BaseDBAsyncClient):
         self.pragmas = kwargs.copy()
         self.pragmas.pop("connection_name", None)
         self.pragmas.pop("fetch_inserted", None)
+        self.pragmas["foreign_keys"] = "ON"
 
         self._connection: Optional[aiosqlite.Connection] = None
         self._lock = asyncio.Lock()